[Refactor: 코드리뷰 후 리팩토링 진행중]
