+++ src/pages/haesoo/Main/FeedContents/FeedContent.scss
@@ -1,4 +1,4 @@
-.main-left {
+.main-left-haesoo {
   display: flex;
   width: 935px;
   margin: 0px 100px;
