+++ src/pages/haesoo/Main/Main.scss
@@ -11,7 +11,7 @@ button {
   height: auto;
   background-color: #fafafa;
 
-  .navigation {
+  .navigation-haesoo {
     position: fixed;
     top: 0;
     display: flex;
